Cleanning and Nav2 params
