+++ robots/launch/tiago_localization.launch.py
@@ -33,29 +33,16 @@ def generate_launch_description():
     nav2_launch_dir = os.path.join(nav2_bringup_dir, 'launch')
     package_dir = get_package_share_directory('robots')
 
-    slam = LaunchConfiguration('slam')
     namespace = LaunchConfiguration('namespace')
-    use_namespace = LaunchConfiguration('use_namespace')
     map_yaml_file = LaunchConfiguration('map')
     use_sim_time = LaunchConfiguration('use_sim_time')
     params_file = LaunchConfiguration('params_file')
-    default_bt_xml_filename = LaunchConfiguration('default_bt_xml_filename')
     autostart = LaunchConfiguration('autostart')
 
     declare_namespace_cmd = DeclareLaunchArgument(
         'namespace',
         default_value='',
         description='Top-level namespace')
-
-    declare_use_namespace_cmd = DeclareLaunchArgument(
-        'use_namespace',
-        default_value='false',
-        description='Whether to apply a namespace to the navigation stack')
-
-    declare_slam_cmd = DeclareLaunchArgument(
-        'slam',
-        default_value='False',
-        description='Whether run a SLAM')
 
     declare_map_yaml_cmd = DeclareLaunchArgument(
         'map',
@@ -71,13 +58,6 @@ def generate_launch_description():
         'params_file',
         default_value=os.path.join(package_dir, 'params', 'nav2_params.yaml'),
         description='Full path to the ROS2 parameters file to use for all launched nodes')
-
-    declare_bt_xml_cmd = DeclareLaunchArgument(
-        'default_bt_xml_filename',
-        default_value=os.path.join(
-            get_package_share_directory('nav2_bt_navigator'),
-            'behavior_trees', 'navigate_w_replanning_and_recovery.xml'),
-        description='Full path to the behavior tree xml file to use')
 
     declare_autostart_cmd = DeclareLaunchArgument(
         'autostart', default_value='true',
@@ -104,66 +84,34 @@ def generate_launch_description():
     head2rangefinder = launch_ros.actions.Node(
       package='tf2_ros',
       executable='static_transform_publisher',
-    arguments=['0.0', '0.0', '0.00', '0.0', '0.0', '3.1415', 'head_2_link', 'range-finder'],
+      arguments=['0.0', '0.0', '0.00', '0.0', '0.0', '3.1415', 'head_2_link', 'range-finder'],
       parameters=[{'use_sim_time': use_sim_time}]
     )
-
-    # Rviz node
-    use_rviz = launch.substitutions.LaunchConfiguration('rviz', default=False)
-    rviz_config = os.path.join(get_package_share_directory('robots'), 'resource', 'odometry.rviz')
-    rviz = launch_ros.actions.Node(
-        package='rviz2',
-        executable='rviz2',
-        output='screen',
-        arguments=['--display-config=' + rviz_config],
-        condition=launch.conditions.IfCondition(use_rviz),
-        parameters=[{'use_sim_time': use_sim_time}]
-    )
-
-    pointcloud_xyz = launch_ros.actions.ComposableNodeContainer(
-      name='container',
-      namespace='',
-      package='rclcpp_components',
-      executable='component_container',
-      composable_node_descriptions=[
-          # Driver itself
-          launch_ros.descriptions.ComposableNode(
-              package='depth_image_proc',
-              plugin='depth_image_proc::PointCloudXyzrgbNode',
-              name='point_cloud_xyz_node',
-              remappings=[('rgb/camera_info', '/camera/camera_info'),
-                          ('rgb/image_rect_color', '/camera/image_raw'),
-                          ('depth_registered/image_rect', '/range_finder/image_depth'),
-                          ('points', '/depth_registered/points')],
-              parameters=[{'use_sim_time': use_sim_time}]
-          ),
-      ],
-      parameters=[{'use_sim_time': use_sim_time}],
-      output='screen',
+    baselink2bf = launch_ros.actions.Node(
+      package='tf2_ros',
+      executable='static_transform_publisher',
+      arguments=['0.0', '0.0', '0.00', '0.0', '0.0', '0.0', 'base_link', 'base_footprint'],
+      parameters=[{'use_sim_time': use_sim_time}]
     )
     
     nav2_bringup_cmd = IncludeLaunchDescription(
-      PythonLaunchDescriptionSource(os.path.join(nav2_launch_dir, 'bringup_launch.py')),
+      PythonLaunchDescriptionSource(os.path.join(nav2_launch_dir, 'localization_launch.py')),
       launch_arguments={'namespace': namespace,
-                        'use_namespace': use_namespace,
-                        'slam': slam,
                         'map': map_yaml_file,
                         'use_sim_time': use_sim_time,
                         'params_file': params_file,
-                        'default_bt_xml_filename': default_bt_xml_filename,
                         'autostart': autostart}.items())
 
     ld = LaunchDescription()
 
     ld.add_action(declare_namespace_cmd)
-    ld.add_action(declare_use_namespace_cmd)
-    ld.add_action(declare_slam_cmd)
     ld.add_action(declare_map_yaml_cmd)
     ld.add_action(declare_use_sim_time_cmd)
     ld.add_action(declare_params_file_cmd)
-    ld.add_action(declare_bt_xml_cmd)
     ld.add_action(declare_autostart_cmd)
     
+    ld.add_action(baselink2bf)
+
     ld.add_action(webots)
     ld.add_action(nav2_bringup_cmd)
