+++ robots/resource/tiago.yaml
@@ -7,5 +7,5 @@ webots_driver:
     robot_base_frame: base_link
     synchronization: false
     use_joint_state_publisher: true
-    wheel_distance: 0.45
-    wheel_radius: 0.10
+    wheel_distance: 0.435
+    wheel_radius: 0.098
